Extract `ConfigurationCacheValueSourceIntegrationTest`
